+++ tests/test_model_methods.py
@@ -5,6 +5,7 @@ from os.path import join, dirname, abspath
 import numpy as np
 import warnings
 import pdb
+import pandas as pd
 import xarray as xr
 from hydromt_wflow.wflow import WflowModel
 import pandas as pd
@@ -56,6 +57,70 @@ def test_setup_staticmaps():
             reproject_method="average",
             wflow_variables=["input.vertical.altitude"],
         )
+
+
+def test_setup_lake(tmpdir):
+    logger = logging.getLogger(__name__)
+    # read model from examples folder
+    root = join(EXAMPLEDIR, "wflow_piave_subbasin")
+
+    # Initialize model and read results
+    mod = WflowModel(root=root, mode="r", data_libs="artifact_data", logger=logger)
+
+    # Create dummy lake rating curves
+    lakes = mod.staticgeoms["lakes"]
+    lake_id = lakes["waterbody_id"].iloc[0]
+    area = lakes["LakeArea"].iloc[0]
+    dis = lakes["LakeAvgOut"].iloc[0]
+    lvl = lakes["LakeAvgLevel"].iloc[0]
+    elev = lakes["Elevation"].iloc[0]
+    lvls = np.linspace(0, lvl)
+
+    df = pd.DataFrame(data={"elevtn": (lvls + elev), "volume": (lvls * area)})
+    df = df.join(
+        pd.DataFrame(
+            {"elevtn": (lvls[-5:-1] + elev), "discharge": np.linspace(0, dis, num=4)}
+        ).set_index("elevtn"),
+        on="elevtn",
+    )
+    fn_lake = join(tmpdir, f"rating_curve_{lake_id}.csv")
+    df.to_csv(fn_lake, sep=",", index=False, header=True)
+
+    # Register as new data source
+    mod.data_catalog.from_dict(
+        {
+            "lake_rating_test_{index}": {
+                "data_type": "DataFrame",
+                "driver": "csv",
+                "path": join(tmpdir, "rating_curve_{index}.csv"),
+                "placeholders": {
+                    "index": [str(lake_id)],
+                },
+            }
+        }
+    )
+    # Update model with it
+    mod.setup_lakes(
+        lakes_fn="hydro_lakes",
+        rating_curve_fns=[f"lake_rating_test_{lake_id}"],
+        min_area=5,
+        add_maxstorage=True,
+    )
+
+    assert f"lake_sh_{lake_id}" in mod.tables
+    assert f"lake_hq_{lake_id}" in mod.tables
+    assert 2 in np.unique(mod.staticmaps["LakeStorFunc"].values)
+    assert 1 in np.unique(mod.staticmaps["LakeOutflowFunc"].values)
+    assert "LakeMaxStorage" not in mod.staticmaps  # no Vol_max column in hydro_lakes
+
+    # Write and read back
+    mod.set_root(join(tmpdir, "wflow_lake_test"))
+    mod.write_tables()
+    test_table = mod.tables[f"lake_sh_{lake_id}"]
+    mod._tables = dict()
+    mod.read_tables()
+
+    assert mod.tables[f"lake_sh_{lake_id}"].equals(test_table)
 
 
 @pytest.mark.timeout(300)  # max 5 min